Update hashing (#49) ~ Basically rename stuff and shorten things
